Add multiplayer support with Socket.IO
