+++ package.json
@@ -1,0 +1,14 @@
+{
+  "name": "ttb-multiplayer",
+  "version": "1.0.0",
+  "description": "Transport Tycoon Basic multiplayer demo",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js",
+    "setup": "npm install"
+  },
+  "dependencies": {
+    "express": "^4.18.2",
+    "socket.io": "^4.7.2"
+  }
+}
