+++ public/index.html
@@ -14,6 +14,7 @@
     <div id="info">
         <span id="money">Money: $0</span>
     </div>
-    <script src="game.js"></script>
+    <script src="/socket.io/socket.io.js"></script>
+    <script src="client.js"></script>
 </body>
 </html>
