+++ setup.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+# Install Node.js dependencies
+npm install
+
+echo "Setup complete. Run 'npm start' to launch the game."
